Refactor Portuguese researcher
