+++ packages/yoastseo/src/languageProcessing/languages/pt/Researcher.js
@@ -1,22 +1,18 @@
-import matchKeywordInSubheadings from "../../researches/base/matchKeywordInSubheadings.js";
-import getProminentWordsForInsights from "../../researches/base/getProminentWordsForInsights";
-import getProminentWordsForInternalLinking from "../../researches/base/getProminentWordsForInternalLinking";
-import getWordForms from "../../researches/_todo/getWordForms";
-import findKeywordInPageTitle from "../../researches/base/findKeywordInPageTitle";
-import { keyphraseDistributionResearcher as keyphraseDistribution } from "../../researches/base/keyphraseDistribution";
-
 import AbstractResearcher from "../../AbstractResearcher";
 
-import getLinkStatistics from "./researches/getLinkStatistics";
-import passiveVoice from "./researches/getPassiveVoice.js";
-import stopWordsInKeyword from "./researches/stopWordsInKeyword";
-import stopWordsInUrl from "./researches/stopWordsInUrl";
-import calculateFleschReading from "./researches/calculateFleschReading";
-import findTransitionWords from "./researches/findTransitionWords";
-import functionWordsInKeyphrase from "./researches/functionWordsInKeyphrase";
-import getSentenceBeginnings from "./researches/getSentenceBeginnings.js";
-import keyphraseLength from "./researches/keyphraseLength";
-import getKeywordDensity from "./researches/getKeywordDensity";
+// All config
+import firstWordExceptions from "./config/firstWordExceptions";
+import functionWords from "./config/functionWords";
+import stopWords from "./config/stopWords";
+import transitionWords from "./config/transitionWords";
+import twoPartTransitionWords from "./config/twoPartTransitionWords";
+import syllables from "./config/syllables.json";
+
+// All helpers
+import getSentenceParts from "./helpers/getSentenceParts";
+import getStemmer from "./helpers/getStemmer";
+import isPassiveSentencePart from "./helpers/isPassiveSentencePart";
+import fleschReadingScore from "./helpers/calculateFleschReadingScore";
 
 /**
  * The researches contains all the researches
@@ -30,23 +26,22 @@ export default class Researcher extends AbstractResearcher {
 	constructor( paper ) {
 		super( paper );
 
-		Object.assign( this.defaultResearches, {
-			calculateFleschReading: calculateFleschReading,
-			stopWordsInKeyword: stopWordsInKeyword,
-			stopWordsInUrl: stopWordsInUrl,
-			findTransitionWords: findTransitionWords,
-			passiveVoice: passiveVoice,
-			getSentenceBeginnings: getSentenceBeginnings,
-			functionWordsInKeyphrase: functionWordsInKeyphrase,
-			matchKeywordInSubheadings: matchKeywordInSubheadings,
-			getLinkStatistics: getLinkStatistics,
-			keyphraseDistribution: keyphraseDistribution,
-			findKeywordInPageTitle: findKeywordInPageTitle,
-			morphology: getWordForms,
-			prominentWordsForInsights: getProminentWordsForInsights,
-			prominentWordsForInternalLinking: getProminentWordsForInternalLinking,
-			keyphraseLength: keyphraseLength,
-			getKeywordDensity: getKeywordDensity,
+		Object.assign( this.config, {
+			language: "pt",
+			isPeriphrastic: true,
+			firstWordExceptions,
+			functionWords,
+			stopWords,
+			transitionWords,
+			twoPartTransitionWords,
+			syllables,
+		} );
+
+		Object.assign( this.helpers, {
+			getSentenceParts,
+			getStemmer,
+			isPassiveSentencePart,
+			fleschReadingScore,
 		} );
 	}
 }
