+++ packages/yoastseo/src/languageProcessing/languages/pt/helpers/internal/getParticiples.js
@@ -1,15 +1,14 @@
 import { forEach, includes } from "lodash-es";
-import getWords from "../../../helpers/word/getWords";
-import participlesFactory from "../config/passiveVoice/participles";
-import PortugueseParticiple from "../config/passiveVoice/PortugueseParticiple";
-
-const participles = participlesFactory();
+import getWords from "../../../../helpers/word/getWords";
+import participles from "../../config/internal/passiveVoiceParticiples";
+import PortugueseParticiple from "../../values/PortugueseParticiple";
 
 /**
  * Creates participle objects for the participles found in a sentence part.
  *
  * @param {string} sentencePartText The sentence part to find participles in.
  * @param {Array} auxiliaries The list of auxiliaries from the sentence part.
+ *
  * @returns {Array} The list with participle objects.
  */
 export default function getParticiples( sentencePartText, auxiliaries ) {
