+++ packages/yoastseo/src/languageProcessing/languages/pt/helpers/internal/stem.js
@@ -21,7 +21,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
-import { findMatchingEndingInArray } from "../morphoHelpers/findMatchingEndingInArray";
+import { findMatchingEndingInArray } from "../../../../helpers/morphology/findMatchingEndingInArray";
 
 
 /**
