+++ packages/yoastseo/src/languageProcessing/languages/pt/helpers/isPassiveSentencePart.js
@@ -1,4 +1,4 @@
-import getParticiples from "./getParticiples";
+import getParticiples from "./internal/getParticiples";
 import determineSentencePartIsPassive from "../../../helpers/passiveVoice/periphrastic/determineSentencePartIsPassive";
 
 /**
@@ -10,7 +10,7 @@ import determineSentencePartIsPassive from "../../../helpers/passiveVoice/periph
  * @returns {boolean} Returns true if passive, otherwise returns false.
  */
 export default function isPassiveSentencePart( sentencePartText, sentencePartAuxiliaries ) {
-	let participles = getParticiples( sentencePartText, sentencePartAuxiliaries );
+	const participles = getParticiples( sentencePartText, sentencePartAuxiliaries );
 
 	return determineSentencePartIsPassive( participles );
 }
